Make queue manager configurable.
